Fix the vscode-languageserver-* imports
This package was importing code from `vscode-languageserver`. However,
it doesn’t specify `vscode-languageserver` as a dependency. All imports
used, were just re-exports from the `vscode-languageserver-protocol`
package. It was also importing code from the
`vscode-languageserver-types` package.

This changes all imports from `vscode-languageserver` and
`vscode-languageserver-types` to `vscode-languageserver-protocol`
imports, and adds this mising dependency.

Also the `vscode-languageserver-textdocument` dependency was moved into
`devDependencies`, because it’s only used in a test and in the example.